♻ : better use of dependency injection
* extract an interface for CredentialsService
* move VaultCredentialsService to its own class
* remove NoopEncryptionService
