+++ src/main/java/io/gaia_app/encryption/EncryptionService.kt
@@ -1,9 +1,5 @@
 package io.gaia_app.encryption
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
 /**
  * Service for encryption/decryption of data
  */
@@ -12,22 +8,5 @@ interface EncryptionService {
     fun encrypt(plaintext:String): String
 
     fun decrypt(cipherText: String): String
-
-}
-
-@Configuration
-class NoOpEncryptionServiceConfig {
-
-    @Bean
-    @ConditionalOnMissingBean(EncryptionService::class)
-    fun noopEncryptionService() = NoOpEncryptionService()
-
-}
-
-class NoOpEncryptionService:EncryptionService{
-
-    override fun encrypt(plaintext: String) = plaintext
-
-    override fun decrypt(cipherText: String) = cipherText
 
 }
